design-system/requirements.txt
faiss-cpu
sentence-transformers
watchdog
